Added Backpropagation to the model, and fixes.
Added Backpropagation to every layer and activation in the model,
as well as to the Loss class.

Added a namespace to the codebase.

Delete the copy and move operators from the classes.

Added a new class to calculate the softmax and the loss at the
same time. This is to optimize the backpropagation.

Signed-off-by: aronkis <aronkis30@gmail.com>