Added Binary Logistic Regression to the network
Added Sigmoid activation to the model and a Loss Binary
Cross Entropy implementation to let the model do Binary Logistic
Regression.

Modified the main.cc to showcase both the classification and the
binary logistic regression implementations.

Added data points for both training and validation to be used with
the Binary Logistic Regression.

Signed-off-by: aronkis <aronkis30@gmail.com>