Adding a setup script and improving CI.
Added a new script that helps with installing the dependecies for
the repository.

Modified the CMakeLists.txt to that the Eigen library is no longer
needed to be installed on the system, but it will be cloned from
github directly.

Added the options to test builds on different operating system.

Fixed an issues with the main.cc which tried to access not
existing files.

Signed-off-by: Aron Kis <kis.aron@analog.com>