Added unit tests and fixed some minor issues.
Added Google Style unit test to the project to test the core
functionalities of the framework. These can be optionally compiled
and executed from the /test folder after build.

Added the option to run the test in the CI.

Fixed an issues where loading an empty model would cause a segmentation fault.

Fixed and issues where training on empty data would cause a segmentation fault.

Fixed the data paths for the example main function when running the MNIST dataset,
which would be downloaded in the build directory.

Fixed an issue with the Softmax Activation, where the backwards method would pass
the matrices twice making the complexity of the function O(N^2). With the current
solution the complexity of the function is O(N).

Signed-off-by: Aron Kis <kis.aron@analog.com>