WIP: Finished Convolutional layer.
Finished the Convolutional layer, it now can be saved, and loaded.
Fully supporting forward and backward propagation. It trains succesfully.

Added 2 new examples to the main.cc to showcase how to use the new layers.

Modified the Model.h so it will accept LayerBase instead of LayerDense to
the trainable layers.

Signed-off-by: Aron Kis <kis.aron@analog.com>