Fixed Save/Load model issue.
Fixed an issue where the savign of the Batch Normalization layer
was not saved correctly, making the saving of the model incorrect.
When loading this model, the evaluation was off by a large margin.

Now the Batch Normalization layer includes a Get/SetParameters()
function.

Also in the data folder there is trained model that can be loaded.
It was trained in 2 epoch so its accuracy and loss it not the best,
but it is good enough for showcasing.
(Validation Accuracy: 0.8727, Validation Loss: 0.349251)y

Signed-off-by: aronkis <aronkis30@gmail.com>