Added a GNU Radio example. Modified radio model.
Added a new GNU Radio example which can be use to generate 4 types
of modulation to be used on the models.
It can be set to random or chosen via the GUI by the user.

The new radio model supports these modulation types, but has to be
retrained for better accuracy.

Signed-off-by: Aron Kis <kis.aron@analog.com>